QRCodeGeneration/requirements.txt
Pillow==9.2.0
pypng==0.20220715.0
PyQRCode==1.2.1
